+++ src/RcppExports.cpp
@@ -109,6 +109,34 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// delldbeta_pl_i_fun_exact
+arma::vec delldbeta_pl_i_fun_exact(arma::vec beta_pl, double c_pl, arma::vec x_pl_ext_i, double y_i);
+RcppExport SEXP _evinf_delldbeta_pl_i_fun_exact(SEXP beta_plSEXP, SEXP c_plSEXP, SEXP x_pl_ext_iSEXP, SEXP y_iSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::vec >::type beta_pl(beta_plSEXP);
+    Rcpp::traits::input_parameter< double >::type c_pl(c_plSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type x_pl_ext_i(x_pl_ext_iSEXP);
+    Rcpp::traits::input_parameter< double >::type y_i(y_iSEXP);
+    rcpp_result_gen = Rcpp::wrap(delldbeta_pl_i_fun_exact(beta_pl, c_pl, x_pl_ext_i, y_i));
+    return rcpp_result_gen;
+END_RCPP
+}
+// d2elldbeta2_pl_i_fun_exact
+arma::mat d2elldbeta2_pl_i_fun_exact(arma::vec beta_pl, double c_pl, arma::vec x_pl_ext_i, double y_i);
+RcppExport SEXP _evinf_d2elldbeta2_pl_i_fun_exact(SEXP beta_plSEXP, SEXP c_plSEXP, SEXP x_pl_ext_iSEXP, SEXP y_iSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::vec >::type beta_pl(beta_plSEXP);
+    Rcpp::traits::input_parameter< double >::type c_pl(c_plSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type x_pl_ext_i(x_pl_ext_iSEXP);
+    Rcpp::traits::input_parameter< double >::type y_i(y_iSEXP);
+    rcpp_result_gen = Rcpp::wrap(d2elldbeta2_pl_i_fun_exact(beta_pl, c_pl, x_pl_ext_i, y_i));
+    return rcpp_result_gen;
+END_RCPP
+}
 // log_lik_fun
 double log_lik_fun(arma::vec gamma_z, arma::vec gamma_pl, arma::vec beta_nb, double alpha_nb, arma::vec beta_pl, double c_pl, arma::mat x_mult_z_ext, arma::mat x_mult_pl_ext, arma::mat x_nb_ext, arma::mat x_pl_ext, arma::vec y, arma::vec offset_nb);
 RcppExport SEXP _evinf_log_lik_fun(SEXP gamma_zSEXP, SEXP gamma_plSEXP, SEXP beta_nbSEXP, SEXP alpha_nbSEXP, SEXP beta_plSEXP, SEXP c_plSEXP, SEXP x_mult_z_extSEXP, SEXP x_mult_pl_extSEXP, SEXP x_nb_extSEXP, SEXP x_pl_extSEXP, SEXP ySEXP, SEXP offset_nbSEXP) {
@@ -132,8 +160,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // update_bfgs_fun
-List update_bfgs_fun(arma::vec gamma_z_in, arma::vec gamma_pl_in, arma::vec beta_nb_in, double alpha_nb_in, arma::vec beta_pl_in, double c_pl, arma::mat x_mult_z_ext, arma::mat x_mult_pl_ext, arma::mat x_nb_ext, arma::mat x_pl_ext, arma::vec y, double max_upd_par, int no_m_bfgs_steps, arma::vec offset_nb);
-RcppExport SEXP _evinf_update_bfgs_fun(SEXP gamma_z_inSEXP, SEXP gamma_pl_inSEXP, SEXP beta_nb_inSEXP, SEXP alpha_nb_inSEXP, SEXP beta_pl_inSEXP, SEXP c_plSEXP, SEXP x_mult_z_extSEXP, SEXP x_mult_pl_extSEXP, SEXP x_nb_extSEXP, SEXP x_pl_extSEXP, SEXP ySEXP, SEXP max_upd_parSEXP, SEXP no_m_bfgs_stepsSEXP, SEXP offset_nbSEXP) {
+List update_bfgs_fun(arma::vec gamma_z_in, arma::vec gamma_pl_in, arma::vec beta_nb_in, double alpha_nb_in, arma::vec beta_pl_in, double c_pl, arma::mat x_mult_z_ext, arma::mat x_mult_pl_ext, arma::mat x_nb_ext, arma::mat x_pl_ext, arma::vec y, double max_upd_par, int no_m_bfgs_steps, arma::vec offset_nb, bool exact_pl);
+RcppExport SEXP _evinf_update_bfgs_fun(SEXP gamma_z_inSEXP, SEXP gamma_pl_inSEXP, SEXP beta_nb_inSEXP, SEXP alpha_nb_inSEXP, SEXP beta_pl_inSEXP, SEXP c_plSEXP, SEXP x_mult_z_extSEXP, SEXP x_mult_pl_extSEXP, SEXP x_nb_extSEXP, SEXP x_pl_extSEXP, SEXP ySEXP, SEXP max_upd_parSEXP, SEXP no_m_bfgs_stepsSEXP, SEXP offset_nbSEXP, SEXP exact_plSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -151,7 +179,8 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< double >::type max_upd_par(max_upd_parSEXP);
     Rcpp::traits::input_parameter< int >::type no_m_bfgs_steps(no_m_bfgs_stepsSEXP);
     Rcpp::traits::input_parameter< arma::vec >::type offset_nb(offset_nbSEXP);
-    rcpp_result_gen = Rcpp::wrap(update_bfgs_fun(gamma_z_in, gamma_pl_in, beta_nb_in, alpha_nb_in, beta_pl_in, c_pl, x_mult_z_ext, x_mult_pl_ext, x_nb_ext, x_pl_ext, y, max_upd_par, no_m_bfgs_steps, offset_nb));
+    Rcpp::traits::input_parameter< bool >::type exact_pl(exact_plSEXP);
+    rcpp_result_gen = Rcpp::wrap(update_bfgs_fun(gamma_z_in, gamma_pl_in, beta_nb_in, alpha_nb_in, beta_pl_in, c_pl, x_mult_z_ext, x_mult_pl_ext, x_nb_ext, x_pl_ext, y, max_upd_par, no_m_bfgs_steps, offset_nb, exact_pl));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -164,8 +193,10 @@ static const R_CallMethodDef CallEntries[] = {
     {"_evinf_ell_pl_i_fun", (DL_FUNC) &_evinf_ell_pl_i_fun, 4},
     {"_evinf_delldbeta_pl_i_fun_approx", (DL_FUNC) &_evinf_delldbeta_pl_i_fun_approx, 4},
     {"_evinf_d2elldbeta2_pl_i_fun_approx", (DL_FUNC) &_evinf_d2elldbeta2_pl_i_fun_approx, 4},
+    {"_evinf_delldbeta_pl_i_fun_exact", (DL_FUNC) &_evinf_delldbeta_pl_i_fun_exact, 4},
+    {"_evinf_d2elldbeta2_pl_i_fun_exact", (DL_FUNC) &_evinf_d2elldbeta2_pl_i_fun_exact, 4},
     {"_evinf_log_lik_fun", (DL_FUNC) &_evinf_log_lik_fun, 12},
-    {"_evinf_update_bfgs_fun", (DL_FUNC) &_evinf_update_bfgs_fun, 14},
+    {"_evinf_update_bfgs_fun", (DL_FUNC) &_evinf_update_bfgs_fun, 15},
     {NULL, NULL, 0}
 };